Detect lack of tests (#14)
Currently, if the test file is invalid (e.g. mixes tab and space
indentation), no tests will be executed by the runner. However, the
situation is not currently treated as an error, resulting in a green
check in GitHub actions. Because of this, after adding each exercise,
the logs have to be manually validated to confirm that the corresponding
tests were correctly executed.

This PR will make the test suite fail if no tests were executed.